fix: remove redundant hooks in profile actions
